requirements.txt
automationhat==0.1.0
beautifulsoup4==4.6.3
blinker==1.3
blinkt==0.1.2
bs4==0.0.1
buttonshim==0.0.2
cachetools==2.1.0
Cap1xxx==0.1.3
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.3.7
cryptography==1.7.1
dialogflow==0.5.1
dnspython==1.15.0
drumhat==0.1.0
enum34==1.1.6
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==0.12.1
fourletterphat==0.1.0
futures==3.2.0
google-api-core==1.5.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-speech==0.36.0
google-cloud-texttospeech==0.2.0
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.3
gpiozero==1.4.1
grpcio==1.15.0
gTTS==2.0.1
gTTS-token==1.1.2
idna==2.7
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
keyring==10.1
keyrings.alt==1.3
MarkupSafe==0.23
mcpi==0.1.1
microdotphat==0.2.1
mote==0.0.3
motephat==0.0.2
numpy==1.12.1
oauthlib==2.0.1
pafy==0.5.4
pantilthat==0.0.5
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
pycrypto==2.6.1
pydub==0.23.0
pygame==1.9.3
pygobject==3.22.0
pyinotify==0.9.6
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyserial==3.2.1
python-dotenv==0.9.1
python-espeak==0.5
python-vlc==3.0.4106
pytz==2018.5
pyxdg==0.25
rainbowhat==0.1.0
requests==2.20.0
requests-oauthlib==0.7.0
RPi.GPIO==0.6.3
rsa==4.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-emu==1.0
sense-hat==2.2.0
simplejson==3.10.0
six==1.11.0
skywriter==0.0.7
sn3218==1.2.7
spidev==3.3
touchphat==0.0.1
twython==3.4.0
unicornhathd==0.0.3
urllib3==1.24
Werkzeug==0.11.15
